Update chatCensoredList.txt
Since we probably aren't going to get around to providing customizable word filters for Lighthouse until a few other things are out of the way, I decided to quickly add this change to remove the gigantic meme-worthy list of words jvyden came up with. We're currently testing the word filter on Beacon, so I kept 'restitched' and added 'h4h' to continue testing the filter on Beacon. It will need to be applied to level titles, reviews, descriptions, etc sometime in the future. Keep in mind someday this file should be customizable by the instance admin.